+++ src/Dephpug/CommandAdapter.php
@@ -7,6 +7,11 @@ class CommandAdapter {
         return self::convertCommandToDBGp($command, $transactionId);
     }
 
+    public function startsWith($big, $small) {
+        $slen = strlen($small);
+        return $slen === 0 || strncmp($big, $small, $slen) === 0;
+    }
+
     public static function convertCommandToDBGp($command, $transactionId) {
         // Example format: $variable
         if(preg_match('/^\$([\w_]+);?$/', $command, $result)) {