+++ src/Dephpug/CommandAdapter.php
@@ -12,6 +12,11 @@ class CommandAdapter {
         return $slen === 0 || strncmp($big, $small, $slen) === 0;
     }
 
+    public function isStatusStop($responses)
+    {
+        return preg_match('/status=\"stopp(?:ed|ing)\"/', $responses);
+    }
+
     public static function convertCommandToDBGp($command, $transactionId) {
         // Example format: $variable
         if(preg_match('/^\$([\w_]+);?$/', $command, $result)) {