+++ console/DebuggerCommand.php
@@ -23,14 +23,14 @@ class DebuggerCommand extends Command
     protected function execute(InputInterface $_, OutputInterface $output)
     {
         $output->writeln(splashScreen('Debugger'));
-
-        try {
-            while(true) {
+        while(true) {
+            try {
                 DbgpServer::start($output);
+            } catch(ExitProgram $e) {
+                $output->writeln("<fg=red;options=bold>{$e}</>");
+                if($e->getCode() == 2) { continue; }
+                exit(1);
             }
-        } catch(ExitProgram $e) {
-            $output->writeln("<fg=red;options=bold>{$e}</>");
-            exit(1);
         }
     }
 }