+++ CHANGELOG.md
@@ -5,14 +5,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [TODO]
-### Add
+- Implement PSR-0 or PSR-4 to project
 - Command to get info about xdebug
-- Create exception to quit dephpugger
 - In config, add option to get a random port
 - Refactor client DBGp protocol
-- When show type array, show in php format, not json
+- When show type array, show in php format, not xml
 - Better use for Singleton $config
 - Run dephpugger for php scripts in cli
+
+## [Unreleased]
+### Add
+- Class to get variable in phpinfo();
+- Create exception to quit dephpugger
 
 ## [0.0.2]
 ### Changed